rewrite logger pimpl with virtual interface
