+++ libs/logger/CMakeLists.txt
@@ -3,7 +3,9 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-add_library(logger logger.cpp)
+add_library(logger
+  logger.cpp
+  logger_spdlog.cpp)
 target_link_libraries(logger
     spdlog
 )
